Refactor handler into separate file
